requirements.txt
pillow
numpy
flask
flask-httpauth
werkzeug